double-precision serialization is not lossless

+++ src/mlpack/methods/cf/normalization/item_mean_normalization.hpp
@@ -76,9 +76,9 @@ class ItemMeanNormalization
       const size_t item = (size_t) datapoint(1);
       datapoint(2) -= itemMean(item);
       // The algorithm omits rating of zero. If normalized rating equals zero,
-      // it is set to the smallest positive double value.
+      // it is set to the smallest positive float value.
       if (datapoint(2) == 0)
-        datapoint(2) = std::numeric_limits<double>::min();
+        datapoint(2) = std::numeric_limits<float>::min();
     });
   }
 
@@ -114,7 +114,7 @@ class ItemMeanNormalization
       // The algorithm omits rating of zero. If normalized rating equals zero,
       // it is set to the smallest positive double value.
       if (tmp == 0)
-        tmp = std::numeric_limits<double>::min();
+        tmp = std::numeric_limits<float>::min();
 
       *it = tmp;
     }
